Fix build: update plugin-pom version to 2.30
Build fails with exception:

Exception in thread "main" java.lang.ArrayIndexOutOfBoundsException: -1
  at com.thoughtworks.xstream.core.util.OrderRetainingMap.entrySet

This is a bug in xstream that was fixed in 1.4.6:

http://x-stream.github.io/jira/746/

The first Jenkins version with xstream >= 1.4.6 is jenkins-1.557.
This means we could use a plugin-pom with a version as low as 1.557.
However plugin-poms prior to 2.x haven't been decoupled from Jenkins
core yet, which makes testing in different Jenkins versions harder
than it needs to be.

All plugin-poms from the 2.x line require a minimum Jenkins version
of 1.625.3 which isn't much worse than 1.557 and still allows to
support Jenkins 1.x.

We therefore upgrade to a 2.x plugin-pom. There's no difference
in the minimum required Jenkins version between any versions of the
2.x line, so we update to the newest one that doesn't introduce
UpperBoundDeps problems, which is 2.30.

As a consequence this commit also adds matrix-project, matrix-auth
and mailer dependencies that have been extracted from jenkins-core.

This commit also updates the email-ext and m2-extra-steps
dependencies to the lowest possible versions that permit executing
a successful build of the project.