Fix GitLegacyValuesProvider for integration tests
GitLegacyValuesProvider should be used when an earlier version of the
git plugin is installed. This is determined by checking for the
availability of the URIish class in the the org.spearce.jgit.transport
package.

That behavior seems to have been accidentally lost in commit
3d74b529ce3b95a51623cf77393af90cd3eea928 causing the git-1.1.9
integration test to fail.

This commit restores the original behavior.