Fix: stack overflow because of circular view references
Two filters can references other views:

* OtherViewsFilter references one other view explicitly
* UnclassifiedViewsFilters references all other views implicitly

This can lead to circular reference chains where one views needs
to know which items another view contains which in turn relies
on the first view to know which items it contains.

We break this by simply ignoring any view that participates
in such a cycle.