Remove git 1.1.11 integration test
The integration test depends on git plugin 1.1.11 and that depends
on org.eclipse.jgit:org.eclipse.jgit:jar:0.12.1.  That jar file is not
available from Maven Central or from the Eclipse repository.  It appears
that we were relying on the cached copy stored on repo.jenkins-ci.org.

The cached copy has been removed.  We've removed cached copies from our
artifact repository as part of:

* https://github.com/jenkins-infra/helpdesk/issues/4533

The test is not valuable enough to rely on an artifact that is not
available.  JGit 1.2.0 was released in 2012. Git plugin 1.1.11 was
released in July 2011.  Let's remove the test.

Testing done:

Confirmed that `mvn clean verify` passes after the change.  Rely on
ci.jenkins.io to confirm that the change is sufficient.